+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [8.0.2] - 2024-01-26
+
+### Added
+
+- Submitted edu-sharing files can now be replaced or removed in edit mode
+- Updated german language string for search button to include hyphen before "Repositorium"
+- Query parameter added to repository URL to prevent references to be added as submissions
+
+### Changed
+
+- Landing page in ES-Repository changed from "search" to "workspace"
+
 ## [8.0.1] - 2024-01-14
 
 ### Changed